6.1.8 Вынесет общую логику обновления
